Separate BaseTimeRange with TimeRange. Add more comments.
