+++ lib/listPage.js
@@ -7,7 +7,10 @@ import Router from 'next/router';
 import url from 'url';
 
 /**
- * @param {object} urlQuery
+ * Go to query with pagination params deleted.
+ *
+ * @param {{[string]: string|undefined}} urlQuery - the new query object to set to URL.
+ *   Set to undefined to delete a key from URL query.
  */
 export function goToUrlQueryAndResetPagination(urlQuery) {
   delete urlQuery.after;
